requirements.txt
https://github.com/thiagopena/geraldo/archive/master.zip
https://github.com/thiagopena/PySIGNFe/archive/master.zip
django==2.0.5
dj-database-url==0.5.*
python-decouple==3.1
pytz==2018.3
reportlab==3.4.*
six==1.11.*
cryptography>=2.3
lxml==4.2.*
pyOpenSSL==17.5.*
signxml==2.5.*
ofxparse==0.20
requests==2.21.0
python-dateutil
django-dbbackup==3.2
django_pdfkit
django-tables2
django-tables2-column-shifter
tablib
django-filter
django-bootstrap3
django-tenants